Minor cleanup of godocs
